refactor rmq
